requirements.txt
faster_whisper==0.6.0
langchain==0.0.218
python-dotenv==1.0.0
pytube==15.0.0
